requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
torchsummary>=1.5.1

# Experiment tracking & HPO
mlflow>=2.10.0
optuna>=3.5.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
aiofiles>=23.0.0
pydantic>=2.0.0

# ONNX
onnx>=1.16.0
onnxruntime>=1.17.0

# Caching
redis>=5.0.0

# Utilities
pyyaml>=6.0
numpy>=1.24.0

# Testing
pytest>=8.0.0
httpx>=0.27.0
